==> analysis/4_Illusion2.qmd <==
---
title: "Illusion Game Reliability"
subtitle: "Test-retest reliability"
date: "`r Sys.Date()`"
editor: source
editor_options: 
  chunk_output_type: console
format:
  html:
    code-fold: true
    self-contained: false
    toc: true
execute: 
  cache: false
---


```{r}
#| echo: false

options(
  digits = 3,
  mc.cores = 4,
  brms.algorithm = "sampling",
  brms.backend = "cmdstanr",
  dplyr.summarise.inform = FALSE,
  knitr.kable.NA = ""
)

fig.width <- 7 * sqrt(2)  # see::golden_ratio(7)
fig.height <- 7

knitr::opts_chunk$set(
  collapse = TRUE,
  dpi = 450,
  fig.width = fig.width,
  fig.height = fig.height
)
```


```{r}
#| message: false
#| warning: false

library(tidyverse)
library(ggdist)
library(ggside)
library(easystats)
library(patchwork)
library(brms)

logmod <- function(x) sign(x) * log(1 + abs(x))
sqrtmod <- function(x) sign(x) * sqrt(abs(x))
cbrtmod <- function(x) sign(x) * (abs(x)**(1 / 3))

df <- read.csv("../data/preprocessed_illusion1.csv") |>
  mutate(
    Block = as.factor(Block),
    Illusion_Side = as.factor(Illusion_Side)
  )
```

```{r}
# Retrieve get_scores() function
knitr::read_chunk("3_Illusion1.qmd", labels="get_scores", from=992, to=1020)
eval(parse(text = knitr::knit_code$get()$get_scores))
knitr::read_chunk("3_Illusion1.qmd", labels="tidy_scores", from=1119, to=1127)
eval(parse(text = knitr::knit_code$get()$tidy_scores))

load("models/illusion1_ebbinghaus_err.Rdata")
load("models/illusion1_ebbinghaus_rt.Rdata")
load("models/illusion1_mullerlyer_err.Rdata")
load("models/illusion1_mullerlyer_rt.Rdata")
load("models/illusion1_verticalhorizontal_err.Rdata")
load("models/illusion1_verticalhorizontal_rt.Rdata")
```


## Individual models

### Computation

```{r}
#| eval: false

get_params <- function(model, data, prior_scale = 1) {
  family <- insight::find_response(model)
  f <- insight::find_formula(model)
  if(family == "RT") {
    formula <- brms::bf(f$conditional, sigma ~ 1, beta ~ 1, family="exgaussian")
  } else {
    formula <- f$conditional
  }
  priors <- bayestestR::model_to_priors(model, 
                                        scale_multiply = prior_scale) |> 
    filter(class %in% c("b", "Intercept") & dpar == "")
  priors <- brms::validate_prior(priors, 
                                 formula, 
                                 data=data)
  s <- capture.output(suppressMessages(
    m <- update(model, 
                formula=formula, 
                newdata=data, prior=priors, 
                refresh=0, 
                silent=0, 
                init=unlist(ifelse(family == "RT", list(0), list(NULL))))
  ))
  parameters::parameters(m) |> 
    filter(str_detect(Parameter, "Illusion_EffectIncongruent:")) |> 
    mutate(Parameter = case_when(
      str_detect(Parameter, "Illusion_Difference:") ~ "Interaction",
      str_detect(Parameter, "_Strength") ~ "Strength",
      TRUE ~ "Diff"
    )) |> 
    select(Parameter, Median) |> 
    pivot_wider(names_from = "Parameter", values_from = "Median") |> 
    data_addsuffix(paste0("_", family))
}


# out <- data.frame()
# already_done <- c()

out <- read.csv("../data/scores_individual.csv")

already_done <- summarize(group_by(out, Participant), n = n()) |>
  filter(n == 4) |>
  pull(Participant)

out <- filter(out, Participant %in% already_done)

# S110 has no data.
for(sub in setdiff(unique(df$Participant), c(already_done, "S110"))) {
  print(sub)
  for(prior_scale in c(0.5, 1, 2, 4)) {
    print(paste0("  ", prior_scale))
    rez1 <- get_params(
      illusion1_ebbinghaus_err,
      data = filter(df, Illusion_Type == "Ebbinghaus", Participant == sub),
      prior_scale = prior_scale) |> 
      data_addprefix("Ind_Ebbinghaus_")
    rez2 <- get_params(
      illusion1_ebbinghaus_rt,
      data = filter(df, Illusion_Type == "Ebbinghaus", Participant == sub),
      prior_scale = prior_scale) |> 
      data_addprefix("Ind_Ebbinghaus_")
    rez3 <- get_params(
      illusion1_mullerlyer_err,
      data = filter(df, Illusion_Type == "MullerLyer", Participant == sub),
      prior_scale = prior_scale) |> 
      data_addprefix("Ind_MullerLyer_")
    rez4 <- get_params(
      illusion1_mullerlyer_rt,
      data = filter(df, Illusion_Type == "MullerLyer", Participant == sub),
      prior_scale = prior_scale) |> 
      data_addprefix("Ind_MullerLyer_")
    rez5 <- get_params(
      illusion1_verticalhorizontal_err,
      data = filter(df, Illusion_Type == "VerticalHorizontal", Participant == sub),
      prior_scale = prior_scale) |> 
      data_addprefix("Ind_VerticalHorizontal_")
    rez6 <- get_params(
      illusion1_verticalhorizontal_rt,
      data = filter(df, Illusion_Type == "VerticalHorizontal", Participant == sub),
      prior_scale = prior_scale) |> 
      data_addprefix("Ind_VerticalHorizontal_")
    
    rez <- cbind(rez1, rez2, rez3, rez4, rez5, rez6)
    rez$Participant <- sub
    rez$PriorScale <- prior_scale
    out <- rbind(out, rez)
    write.csv(out, "../data/scores_individual.csv", row.names = FALSE)
  }
}
```

### Comparison

#### Correlation

```{r}
data <- read.csv("../data/scores_individual.csv") |> 
  # standardize(exclude = c("Participant", "PriorScale")) |> 
  data_to_long(-starts_with("Participant|PriorScale"),
               values_to = "IndividualModel",
               names_to = "Score") |>
  mutate(Score = str_remove(Score, "Ind_")) |> 
  full_join(
    read.csv("../data/scores_illusion1.csv") |> 
      # standardize(exclude = c("Participant")) |> 
      data_to_long(-Participant, values_to = "MixedModel", names_to = "Score"),
    by = c("Participant", "Score")) |> 
  mutate(PriorScale = as.factor(PriorScale))
   

data1 <- read.csv("../data/scores_individual.csv") |> 
  filter(PriorScale == 4) |> 
  full_join(datawizard::data_addprefix(read.csv("../data/scores_illusion1.csv"), "Mod_", exclude = "Participant"), by = "Participant")

  
correlation(select(data1, starts_with("Ind_")), select(data1, starts_with("Mod_"))) |> 
  mutate(Parameter1 = str_remove_all(Parameter1, "Ind_"),
         Parameter2 = str_remove_all(Parameter2, "Mod_")) |> 
  filter(Parameter1 == Parameter2) |> 
  separate(Parameter1, into = c("Illusion", "Parameter"), sep = "_", extra="merge") |>
  select(-Parameter2, -Method, -n_Obs) |> 
  arrange(Illusion, desc(r)) |>
  insight::format_table() |> 
  gt::gt() |> 
  gt::opt_stylize() |> 
  gt::tab_header("Individual Model vs. Mixed Model Correlation")
```

#### Effect of Priors

```{r}
model <- lm(MixedModel ~ Score / (IndividualModel * PriorScale), data=data)
slopes <- estimate_slopes(model, trend = "IndividualModel", by = c("Score", "PriorScale"))

slopes |> 
  cbind(
    effectsize::t_to_r(slopes$t, slopes$df_error) |> 
      datawizard::data_addprefix("r_", exclude = "r")) |> 
  ggplot(aes(x = PriorScale, y = r)) +
  geom_line(aes(color = Score, group=Score), position = position_dodge(width = 0.2)) +
  geom_pointrange(aes(ymin = r_CI_low, ymax = r_CI_high, color = Score), 
                  position = position_dodge(width = 0.2)) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  theme_modern()
```


## Test-Retest

### Session 2 Scores

```{r}
df2 <- read.csv("../data/preprocessed_illusion2.csv") |>
  mutate(
    Block = as.factor(Block),
    Illusion_Side = as.factor(Illusion_Side),
    Illusion_Effect = fct_relevel(as.factor(Illusion_Effect), "Incongruent", "Congruent")
  )
```



```{r}
#| eval: false

illusion2_ebbinghaus_err <- update(illusion1_ebbinghaus_err,
                                   newdata=filter(df2, Illusion_Type == "Ebbinghaus"))
save(illusion2_ebbinghaus_err, file = "models/illusion2_ebbinghaus_err.Rdata")

illusion2_ebbinghaus_rt <- update(illusion1_ebbinghaus_rt,
                                  newdata=filter(df2, Illusion_Type == "Ebbinghaus", Error==0))
save(illusion2_ebbinghaus_rt, file = "models/illusion2_ebbinghaus_rt.Rdata")

illusion2_mullerlyer_err <- update(illusion1_mullerlyer_err,
                                  newdata=filter(df2, Illusion_Type == "MullerLyer"))
save(illusion2_mullerlyer_err, file = "models/illusion2_mullerlyer_err.Rdata")

illusion2_mullerlyer_rt <- update(illusion1_mullerlyer_rt,
                                 newdata=filter(df2, Illusion_Type == "MullerLyer", Error==0))
save(illusion2_mullerlyer_rt, file = "models/illusion2_mullerlyer_rt.Rdata")

illusion2_verticalhorizontal_err <- update(illusion1_verticalhorizontal_err,
                                          newdata=filter(df2, Illusion_Type == "VerticalHorizontal"))
save(illusion2_verticalhorizontal_err, file = "models/illusion2_verticalhorizontal_err.Rdata")

illusion2_verticalhorizontal_rt <- update(illusion1_verticalhorizontal_rt,
                                         newdata=filter(df2, Illusion_Type == "VerticalHorizontal", Error==0))
save(illusion2_verticalhorizontal_rt, file = "models/illusion2_verticalhorizontal_rt.Rdata")
```

### Comparison

```{r}
# for(m in c("ebbinghaus", "mullerlyer", "verticalhorizontal")) {
#   for(i in c("err", "rt")) {
#     load(paste0("models/illusion2_", m, "_", i, ".Rdata"))
#   }
# }
load("models/illusion2_ebbinghaus_err.Rdata")
load("models/illusion2_mullerlyer_err.Rdata")
load("models/illusion2_verticalhorizontal_err.Rdata")

scores2 <- tidy_scores(get_scores(illusion2_ebbinghaus_err, "Ebbinghaus")) |> 
  full_join(tidy_scores(get_scores(illusion2_mullerlyer_err, "MullerLyer")), by = "Participant") |> 
  full_join(tidy_scores(get_scores(illusion2_verticalhorizontal_err, "VerticalHorizontal")), by = "Participant") 
```

```{r}
scores <- scores2 |> 
  datawizard::data_addsuffix("_S2", exclude="Participant") |> 
  dplyr::full_join(read.csv("../data/scores_illusion1.csv"), by = "Participant")

correlation(select(scores, -ends_with("_S2")), select(scores, ends_with("_S2"))) |> 
  mutate(Parameter2 = str_remove(Parameter2, "_S2")) |>
  filter(Parameter1 == Parameter2) |> 
  separate(Parameter1, into = c("Illusion", "Parameter"), sep = "_", extra="merge") |>
  select(-Parameter2, -Method, -n_Obs) |> 
  arrange(Illusion, desc(r)) |>
  insight::format_table() |> 
  gt::gt() |> 
  gt::opt_stylize() |> 
  gt::tab_header("Test-Retest Correlations")
```
